Move embeddings to contrib module per reviewer request
