+++ python/cudaq/__init__.py
@@ -381,8 +381,8 @@ _LAZY_ATTRS = {
     'evolve': '.dynamics.evolution',
     'evolve_async': '.dynamics.evolution',
     'IntermediateResultSave': '.dynamics.helpers',
-    'amplitude_encode': '.kernels.embeddings',       # <-- ADD
-    'angular_encode': '.kernels.embeddings',          # <-- ADD
+    'amplitude_encode': '.contrib.embeddings',       # <-- ADD
+    'angular_encode': '.contrib.embeddings',          # <-- ADD
 }
 
 _LAZY_SUBMODULES = {
