+++ python/cudaq/contrib/__init__.py
@@ -7,3 +7,4 @@
 # ============================================================================ #
 
 from .qiskit_convert import from_qiskit, from_qasm
+from .embeddings import amplitude_encode, angular_encode
